feat: limit for legal usage
